+++ Diagnostics/Assets/Basic/LDL/Haptics/LDLHapticsController.cs
@@ -210,7 +210,7 @@ public class LDLHapticsController : MonoBehaviour, IRemoteControllable
         else
         {
             Debug.Log("LDL: Starting new measurement");
-            HTS_Server.SendMessage(_mySceneName, "Status:Staring new measurement");
+            HTS_Server.SendMessage(_mySceneName, "Status:Starting new measurement");
 
             File.Delete(_stateFile);
             StartMeasurement();